google-app-engine/requirements.txt
pandas==1.2.1
dash==1.19.0
dash-html-components==1.1.2
dash-core-components==1.15.0
dash-table==4.11.2
dash-daq==0.5.0
gunicorn==20.0.4
google-cloud-firestore>=2.0.2